+++ src/Products/ZCatalog/plan.py
@@ -25,6 +25,7 @@ from Acquisition import aq_parent
 from zope.dottedname.resolve import resolve
 
 from Products.PluginIndexes.interfaces import IDateRangeIndex
+from Products.PluginIndexes.interfaces import ILimitedResultIndex
 from Products.PluginIndexes.interfaces import IUniqueValueIndex
 
 
@@ -301,10 +302,20 @@ class CatalogPlan(object):
         self.end_time = time.time()
         self.duration = self.end_time - self.start_time
         # Make absolutely sure we never omit query keys from the plan
+        current = PriorityMap.get_entry(self.cid, self.key)
         for key in self.query.keys():
             key = self.querykey_to_index.get(key, key)
             if key not in self.benchmark.keys():
-                self.benchmark[key] = Benchmark(0, 0, False)
+                if current and key in current:
+                    self.benchmark[key] = Benchmark(*current[key])
+                else:
+                    if key in self.catalog.indexes:
+                        index = self.catalog.indexes[key]
+                        self.benchmark[key] = Benchmark(
+                            0, 0, ILimitedResultIndex.providedBy(index)
+                        )
+                    else:
+                        self.benchmark[key] = Benchmark(0, 0, False)
         PriorityMap.set_entry(self.cid, self.key, self.benchmark)
         self.log()